+++ skillfarm/tests/managers/test_characterskill_manager.py
@@ -46,7 +46,9 @@ class TestCharacterSkillManager(SkillFarmTestCase):
                 "unallocated_sp": 250000,
             },
         )
-        self.skillfarm_audit.update_skills(force_refresh=False)
+        self.skillfarm_audit.skillfarm_skills.update_or_create_esi(
+            character=self.skillfarm_audit, force_refresh=False
+        )
 
         self.assertSetEqual(
             set(